Move notifier into usecase package
